+++ MyLayout/ViewController.m
@@ -34,6 +34,7 @@
 #import "FLLTest2ViewController.h"
 #import "FLLTest3ViewController.h"
 #import "FLLTest4ViewController.h"
+#import "FLLTest5ViewController.h"
 
 
 #import "AllTest1ViewController.h"
@@ -134,6 +135,9 @@
                            },
                          @{@"title":NSLocalizedString(@"4.FlowLayout - Weight", @""),
                            @"class":[FLLTest4ViewController class]
+                           },
+                         @{@"title":NSLocalizedString(@"5.FlowLayout - Paging", @""),
+                           @"class":[FLLTest5ViewController class]
                            },
                          @{@"title":NSLocalizedString(@"1.FloatLayout - Float", @""),
                            @"class":[FOLTest1ViewController class]